fix: Add retry logic to backfill approval, make server type an enum
The backfill approval endpoint can return 429 (Too Many Requests) - retry the request once in this case to attempt immediate approval. If this doesn't work, we do not retry further, except for the fact a new request will be made a second later. Also, move server types to an iota-based enum.