feat(a2s): Support S2C_CHALLENGE
Support the `S2C_CHALLENGE` response to prevent a reflection amplification attack as defined in the documentation: https://developer.valvesoftware.com/wiki/Server_queries#Request_Format. Part of this involved refactoring the query handling to make it more generic. Additionally, allocate `queryReadBufferSizeBytes` for the query buffer, so we can read the maximum amount of data configured to be read in the query handler.